requirements.txt
Flask==2.3.3
PyJWT==2.8.0
Werkzeug==2.3.7

# Core dependencies
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-chroma>=0.0.10
chromadb>=0.4.18
openai>=1.0.0

# Document processing
pypdf>=3.15.1
sentence-transformers>=2.2.2

# Web scraping and data collection
requests>=2.31.0
beautifulsoup4>=4.12.2

# HTTP client for banking API
httpx>=0.25.0

# Security
cryptography>=41.0.0

# MCP framework
mcp==1.9.1
fastapi>=0.100.0
uvicorn>=0.22.0
aiohttp>=3.8.5
asyncio>=3.4.3

# Database
# sqlite3 is part of the Python standard library

# Optional for older Python versions
dataclasses; python_version<"3.7"
